// web/src/app/core/api/admin-generated/models/content-batch-item-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ContentBatchImportAction } from '../models/content-batch-import-action';
import { ContentBatchItemStatus } from '../models/content-batch-item-status';
export interface ContentBatchItemResponse {
  action?: ContentBatchImportAction;
  catalogPath?: string;
  contentHash?: (string | null);
  id?: string;
  ordinal?: (number | string);
  problemId?: (number | string | null);
  revisionId?: (string | null);
  safeFailureCategory?: (string | null);
  safeFailureMessage?: (string | null);
  slug?: string;
  status?: ContentBatchItemStatus;
  title?: string;
  updatedAt?: string;
}